Fix android studio crash
