+++ mvicore-plugin/idea/build.gradle
@@ -1,5 +1,5 @@
 plugins {
-    id "org.jetbrains.intellij" version "0.4.7"
+    id "org.jetbrains.intellij" version "0.4.9"
 }
 
 apply plugin: 'java'
@@ -13,7 +13,7 @@ version = '0.0.1'
 
 intellij {
     pluginName 'mvicore-plugin'
-    version "2019.1"
+    version "2019.1.3"
     plugins = ['android']
 }
